arm: dts: rk3128h-box-avb: Add the possible "boot devices"
During uevent processing, some "by-name" symlinks will be created.
The following symlinks will then be created on the device.

	/dev/block/by-name/<partition> -> /dev/block/<type>/<device>/by-name/<partition>

Note that both <type> and <device> are skipped in the newly create symlinks.
It assumes there is no more than one devices with the same <partition>,
which is the assumption of current first stage mount flow.

When 'boot_devices' in DT is absent, it fallbacks to extract
'boot_devices' from fstab settings.

Change-Id: Ie7f721b1ae53245c324f7ef2c707e7776aeacf66
Signed-off-by: Zhangbin Tong <zebulun.tong@rock-chips.com>